requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.1.1
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake
python-dotenv
pandas